・区切り線を追加 ・不必要なスクロールビューを削除 ・各UI項目を別々の関数に分離 ・複数インスタンス編集でDynamicPathの要素数が増えたとき、意図していないパスが代入されていたのを修正 ・MultipleEditorの設定項目を追加
