+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c8bac79a87224394ea56863947a1c7a0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
