+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript/GUIElement_PackagePrefsElementList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bbc81eada0dc9aa42918755ff8bd21fe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
