+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/UnityPackageExporterEditor.cs
@@ -11,7 +11,6 @@ namespace MizoreNekoyanagi.PublishUtil.PackageExporter
     {
         public static string HelpBoxText;
         public static MessageType HelpBoxMessageType;
-        public static Vector2 scroll;
         public static string variable_key_temp;
         public MizoresPackageExporter t;
         private void OnEnable( ) {
